+++ src/OrderedList.php
@@ -31,7 +31,7 @@ final class OrderedList implements MemberList
      *
      * @return static
      */
-    public static function from(InnerList|Item|ByteSequence|Token|bool|int|float|string ...$members): self
+    public static function new(InnerList|Item|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
     {
         return self::fromList($members);
     }
@@ -39,7 +39,7 @@ final class OrderedList implements MemberList
     /**
      * @param iterable<InnerList<int, Item>|Item|DataType> $members
      */
-    public static function fromList(iterable $members = []): self
+    public static function fromList(iterable $members): self
     {
         $instance = new self();
         foreach ($members as $member) {
@@ -49,7 +49,7 @@ final class OrderedList implements MemberList
         return $instance;
     }
 
-    private static function filterMember(StructuredField|ByteSequence|Token|bool|int|float|string $member): InnerList|Item
+    private static function filterMember(StructuredField|ByteSequence|Token|Stringable|bool|int|float|string $member): InnerList|Item
     {
         return match (true) {
             $member instanceof InnerList, $member instanceof Item => $member,
@@ -65,12 +65,11 @@ final class OrderedList implements MemberList
      */
     public static function fromHttpValue(Stringable|string $httpValue): self
     {
-        $instance = new self();
-        foreach (Parser::parseList($httpValue) as $value) {
-            $instance->push(is_array($value) ? InnerList::fromList(...$value) : $value);
-        }
-
-        return $instance;
+        return self::new()
+            ->push(...array_map(
+                fn ($value) => is_array($value) ? InnerList::fromList(...$value) : $value,
+                Parser::parseList($httpValue)
+            ));
     }
 
     public function toHttpValue(): string
@@ -136,7 +135,7 @@ final class OrderedList implements MemberList
      *
      * @param InnerList<int, Item>|Item|DataType ...$members
      */
-    public function unshift(StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
+    public function unshift(StructuredField|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
     {
         $this->members = [...array_map(self::filterMember(...), array_values($members)), ...$this->members];
 
@@ -148,7 +147,7 @@ final class OrderedList implements MemberList
      *
      * @param InnerList<int, Item>|Item|DataType ...$members
      */
-    public function push(StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
+    public function push(StructuredField|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
     {
         $this->members = [...$this->members, ...array_map(self::filterMember(...), array_values($members))];
 
@@ -162,7 +161,7 @@ final class OrderedList implements MemberList
      *
      * @throws InvalidOffset If the index does not exist
      */
-    public function insert(int $index, StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
+    public function insert(int $index, StructuredField|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
     {
         $offset = $this->filterIndex($index);
         match (true) {
@@ -182,7 +181,7 @@ final class OrderedList implements MemberList
      *
      * @throws InvalidOffset If the index does not exist
      */
-    public function replace(int $index, StructuredField|ByteSequence|Token|bool|int|float|string $member): self
+    public function replace(int $index, StructuredField|ByteSequence|Token|Stringable|bool|int|float|string $member): self
     {
         if (null === ($offset = $this->filterIndex($index))) {
             throw InvalidOffset::dueToIndexNotFound($index);